access_gnomevfs: Read function take a size_t and not an int.

+++ modules/access/gnomevfs.c
@@ -65,7 +65,7 @@ vlc_module_end ()
  * Exported prototypes
  *****************************************************************************/
 static int  Seek( access_t *, int64_t );
-static int  Read( access_t *, uint8_t *, int );
+static int  Read( access_t *, uint8_t *, size_t );
 static int  Control( access_t *, int, va_list );
 
 struct access_sys_t
@@ -285,7 +285,7 @@ static void Close( vlc_object_t * p_this )
 /*****************************************************************************
  * Read: standard read on a file descriptor.
  *****************************************************************************/
-static int Read( access_t *p_access, uint8_t *p_buffer, int i_len )
+static int Read( access_t *p_access, uint8_t *p_buffer, size_t i_len )
 {
     access_sys_t *p_sys = p_access->p_sys;
     GnomeVFSFileSize i_read_len;
